+++ R/dbi-result.R
@@ -43,7 +43,8 @@ BigQueryResult <- function(conn, sql, ...) {
     page_size = conn@page_size,
     quiet = conn@quiet,
     cursor = cursor(nrow),
-    bigint = conn@bigint
+    bigint = conn@bigint,
+    billing = conn@billing
   )
   res
 }
@@ -61,7 +62,8 @@ setClass(
     page_size = "numeric",
     quiet = "logical",
     cursor = "list",
-    bigint = "character"
+    bigint = "character",
+    billing = "character"
   )
 )