+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * `bq_stream()` allows streaming data into BigQuery table (@byapparov, #383)
 
+* Add `billing` slot to `BigQueryResult`.
+
+
 # bigrquery 1.3.2
 
 * BigQuery `BYTES` and `GEOGRAPHY` column types are now supported via